Got the tests to compile and run. Coverage is at 83% right now.
Changed the `u8casecmp()` to use lowercase characters for the compare.
Changed the `wctombs()` to return -1 on errors.
Fixed the `mbstowc()` skip function, it was not skip the 0xBF bytes.
Fixed the name of the test binary in mk.
Removed the tools folder from the coverage for now. It's non-existant.
Fixed the `doc/index.php` and `glob()` the libutf9 directory, not `advgetopt`.