Got the tests to 100% coverage again.
Renamed the tests without the "unittest_" introducer. It's not useful since we only have one unit test anyway.
Added an start_with_bom() function to check whether a buffer starts with a BOM character.
Fixed the various standalone characters that are u32char_t to start with a U instead of an L (i.e. U'\0').