Started implementing Unicode support.
Unicode allows us to canonicalize UTF-8 strings which is important for a valid strcmp().
Cleaned up (simplified) license blurp.
Added SnapDev as a dependency so we can easily implement the Unicode parser.
Added a tool to run the parser (which is part of the library).
Updated the exception declarations with our macros.