Enhanced the iterator to better match the std lib.
The iterator now derive from std::iterator as expected.
Made the iterator copyable (i.e. no const / const reference).
Added the operator - () between two libutf8 iterators.
Fixed a case in the catch_iterator test which would always fail an if() statement.
Added a few checks to make sure that the iterators are working as expected.
Among the iterator tests, added a couple of rewind() calls to also verify that.