Added tests to get more coverage.
Verify the UTF-16 length and validation functions.
Added a test of the output of a char32_t to an iostream.
Added a function to convert a char32_t to UTF-16--if surrogate are required, convert using those.
Added += operators to the std::string with char32_t (and int/unsigned for cases where char was not used and instead a value like 0x40 is viewed as unsigned).
Cleaned up sections in the string validation test.
Moved tests expected to throw in a separate test marked as "[invalid]".